requirements.txt
streamlit
moviepy
openai-whisper
torch
pyannote.audio
numpy
pillow
transformers
python-dotenv
openai